+++ CHANGELOG.md
@@ -10,6 +10,7 @@ x.y.z Release notes (yyyy-MM-dd)
 
 * Change default authentication mode to `AuthenticationMode.NATIVE`
 * Request with multiple accept headers correctly generated e.g. `GetTermsAPI`
+* Remove `autoClose` option when starting the `AppActivity`
 
 ### Enhancements
 
@@ -23,6 +24,7 @@ x.y.z Release notes (yyyy-MM-dd)
 ### Internal
 
 * Use gas station endpoint as fallback for POI position if none was found in the database when observing a single POI by ID
+* Return location based apps based on `appsDistanceThresholdInMeters` configuration and not if the location is in the polygon
 
 
 8.2.0 Release notes (2021-07-12)